ingest/requirements.txt
requests==2.32.3
psycopg[binary]==3.2.6
python-dateutil==2.9.0.post0
